+++ Data/Models/FoodJournal/FoodItem.cs
@@ -10,13 +10,13 @@ public class FoodItem
 
     public string Units { get; set; } = "";
 
-    public int Calories { get; set; }
+    public double Calories { get; set; }
 
-    public int Carbs { get; set; }
+    public double Carbs { get; set; }
 
-    public int Fats { get; set; }
+    public double Fats { get; set; }
 
-    public int Proteins { get; set; }
+    public double Proteins { get; set; }
 
     // Nav props
     public virtual List<MealFoodServing> MealFoodServings { get; set; } = [];
