+++ Data/Models/FoodJournal/UserGoals.cs
@@ -4,13 +4,13 @@ public class UserGoals
 {
     public int Id { get; set; }
 
-    public int Calories { get; set; }
+    public double Calories { get; set; }
 
-    public int Carbs { get; set; }
+    public double Carbs { get; set; }
 
-    public int Fats { get; set; }
+    public double Fats { get; set; }
 
-    public int Protein { get; set; }
+    public double Protein { get; set; }
 
     public ApplicationUser User { get; set; } = null!;
 }
